% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adonis_rarperm_pq.R
\name{adonis_rarperm_pq}
\alias{adonis_rarperm_pq}
\title{Permanova (adonis) on permutations of rarefaction even depth}
\usage{
adonis_rarperm_pq(
  physeq,
  formula,
  dist_method = "bray",
  merge_sample_by = NULL,
  na_remove = FALSE,
  rarefy_nb_seqs = FALSE,
  verbose = TRUE,
  nperm = 99,
  progress_bar = TRUE,
  quantile_prob = 0.975,
  sample.size = min(phyloseq::sample_sums(physeq)),
  ...
)
}
\arguments{
\item{physeq}{(required) A \code{\link[phyloseq]{phyloseq-class}} object
obtained using the \code{phyloseq} package.}

\item{formula}{(required) the right part of a formula for
\code{\link[vegan:adonis2]{vegan::adonis2()}}. Variables must be present in the \code{physeq@sam_data}
slot.}

\item{dist_method}{(default "bray") the distance used. See
\code{\link[phyloseq:distance]{phyloseq::distance()}} for all available distances or run
\code{\link[phyloseq:distanceMethodList]{phyloseq::distanceMethodList()}}. For aitchison and robust.aitchison
distance, \code{\link[vegan:vegdist]{vegan::vegdist()}} function is directly used.}

\item{merge_sample_by}{a vector to determine which samples to merge using
the \code{\link[MiscMetabar:merge_samples2]{MiscMetabar::merge_samples2()}} function. Need to be in
\code{physeq@sam_data}.}

\item{na_remove}{(logical, default FALSE) If set to TRUE, remove samples with
NA in the variables set in formula.}

\item{rarefy_nb_seqs}{(logical, default FALSE) Rarefy each sample
(before merging if merge_sample_by is set) using
\code{\link[MiscMetabar:rarefy_even_depth_pq]{MiscMetabar::rarefy_even_depth_pq()}} inside \code{\link[MiscMetabar:adonis_pq]{MiscMetabar::adonis_pq()}}.}

\item{verbose}{(logical, default TRUE) If TRUE, prompt some messages.}

\item{nperm}{(int, default 99) The number of permutations to perform.}

\item{progress_bar}{(logical, default TRUE) Do we print progress during
the calculation?}

\item{quantile_prob}{(float, \verb{[0:1]}) the value to compute the quantile.
Minimum quantile is computed using \code{1 - quantile_prob}.}

\item{sample.size}{(int) A single integer value equal to the number of
reads being simulated, also known as the depth. See
\code{\link[phyloseq:rarefy_even_depth]{phyloseq::rarefy_even_depth()}} and \code{\link[MiscMetabar:rarefy_even_depth_pq]{MiscMetabar::rarefy_even_depth_pq()}}.}

\item{...}{Other params to be passed on to \code{\link[MiscMetabar:adonis_pq]{MiscMetabar::adonis_pq()}}
function.}
}
\value{
A list of three dataframe representing the mean, the minimum quantile
and the maximum quantile value for adonis results. See
\code{\link[MiscMetabar:adonis_pq]{MiscMetabar::adonis_pq()}}.
}
\description{
\lifecycle{experimental}

Permanova are computed on a given number of rarefaction with different
seed number. This reduce the risk of a random drawing of an exceptional
situation of an unique rarefaction. Each permutation rarefies \code{physeq} to
even depth (using \code{\link[MiscMetabar:rarefy_even_depth_pq]{MiscMetabar::rarefy_even_depth_pq()}}) and runs
\code{\link[MiscMetabar:adonis_pq]{MiscMetabar::adonis_pq()}}; results are summarized across permutations.
}
\examples{
\donttest{
if (requireNamespace("vegan")) {
  data_fungi_woNA <-
    subset_samples(data_fungi_mini, !is.na(Time) & !is.na(Height))
  adonis_rarperm_pq(data_fungi_woNA, "Time*Height", na_remove = TRUE, nperm = 3)
}
}
}
\seealso{
\code{\link[MiscMetabar:adonis_pq]{MiscMetabar::adonis_pq()}}, \code{\link[=hill_test_rarperm_pq]{hill_test_rarperm_pq()}},
\code{\link[=var_par_rarperm_pq]{var_par_rarperm_pq()}}
}
\author{
Adrien Taudière
}
