% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootpq-package.R
\docType{package}
\name{bootpq-package}
\alias{bootpq}
\alias{bootpq-package}
\title{bootpq: Bootstrapping by Rarefaction Permutations for 'phyloseq' Objects}
\description{
Bootstrapping tools for 'phyloseq' objects based on repeated rarefaction to even sequencing depth. Statistical analyses (PERMANOVA, Hill-number group tests, and variation partitioning) are computed across many rarefaction permutations, each with a different random seed, and summarized to reduce the risk of drawing conclusions from a single, possibly atypical, rarefaction. Part of the 'pqverse' ecosystem and built on top of 'MiscMetabar'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://adrientaudiere.github.io/bootpq/}
  \item \url{https://github.com/adrientaudiere/bootpq}
  \item Report bugs at \url{https://github.com/adrientaudiere/bootpq/issues}
}

}
\author{
\strong{Maintainer}: Adrien Taudière \email{adrien.taudiere@zaclys.net} (\href{https://orcid.org/0000-0003-1088-1182}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Adrien Taudière \email{adrien.taudiere@zaclys.net} (\href{https://orcid.org/0000-0003-1088-1182}{ORCID}) [copyright holder]
}

}
\keyword{internal}
